+++ mitra_models/src/conversations/test_utils.rs
@@ -8,6 +8,7 @@ impl Conversation {
         Self {
             id: Uuid::new_v4(),
             root_id,
+            group_id: None,
             is_managed: true,
             object_id: None,
             audience: Some(AP_PUBLIC.to_owned()),