+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Changed MSRV to 1.81.0.
 - Add actors from `audience` field to post audience.
+- Associate group posts with group actors.
 
 ## [5.4.0] - 2026-05-27